requirements.txt
Flask==2.1.0
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.4.0
pymongo==3.12.0
Werkzeug==2.0.2
python-dotenv==0.19.2
